Utilities

DynamicStructs defines a few functions to simplify working with dynamic properties, as well as the more general case where propertynames(x) != fieldnames(typeof(x)).

DynamicStructs.NoFields — Type

Insert as second argument to Base.propertynames, propertyvalues, or propertypairs to exclude field names.

DynamicStructs.OnlyFields — Type

Insert as second argument to Base.propertynames, propertyvalues, or propertypairs to include only field names.

Base.propertynames — Function

Base.propertynames(x, ::Type{NoFields}, private=false)

Excludes field names.

Base.propertynames(x, ::Type{OnlyFields}, private=false)

Includes only field names.

DynamicStructs.propertyvalues — Function

propertyvalues(x, args...)

Get a tuple of the current property values. args is propagated to Base.propertynames.

DynamicStructs.propertypairs — Function

propertypairs(x, args...)

Get a tuple of the current property names and values. args is propagated to Base.propertynames.

Examples

julia> using DynamicStructs
julia> @dynamic struct Spaceship
           name::String
       end
julia> ship = Spaceship("Blip-A"; pressure=28.0)Main.Spaceship("Blip-A"; pressure = 28.0)
julia> propertynames(ship, OnlyFields)(:name,)
julia> propertynames(ship, NoFields)(:pressure,)
julia> propertyvalues(ship)("Blip-A", 28.0)
julia> propertyvalues(ship, NoFields)(28.0,)
julia> propertypairs(ship)(:name => "Blip-A", :pressure => 28.0)
julia> propertypairs(ship, OnlyFields)(:name => "Blip-A",)
julia> propertypairs(ship, true) # shows all properties, including private(:__properties => DynamicStructs.Properties(pressure = 28.0), :name => "Blip-A", :pressure => 28.0)